Refactor canvas filters to include new time syntax
